+++ src/main/kotlin/me/partlysanestudios/partlysaneskies/render/waypoint/BeamRenderer.kt
@@ -1,9 +1,7 @@
-package me.partlysanestudios.partlysaneskies.render
+package me.partlysanestudios.partlysaneskies.render.waypoint
 
 import me.partlysanestudios.partlysaneskies.render.RenderPrimitives.drawBoxFill
-import me.partlysanestudios.partlysaneskies.render.RenderPrimitives.drawBoxOutline
-import me.partlysanestudios.partlysaneskies.render.points.Point3d
-import me.partlysanestudios.partlysaneskies.utils.ChatUtils
+import me.partlysanestudios.partlysaneskies.utils.vectors.Point3d
 import net.minecraft.client.Minecraft
 import net.minecraft.client.renderer.GlStateManager
 import net.minecraft.client.renderer.Tessellator
@@ -12,9 +10,8 @@ import net.minecraft.util.BlockPos
 import org.lwjgl.opengl.GL11
 import java.awt.Color
 
-
-object BlockHighlightRenderer {
-    fun renderColoredBlockHighlight(pos: BlockPos, outlineColor: Color, fillColor: Color) {
+object BeamRenderer {
+    fun renderBeam(pos: BlockPos, outlineColor: Color, fillColor: Color) {
         val minecraft = Minecraft.getMinecraft()
         val renderManager = minecraft.renderManager
         val tessellator = Tessellator.getInstance()
@@ -32,26 +29,21 @@ object BlockHighlightRenderer {
         GlStateManager.tryBlendFuncSeparate(GL11.GL_SRC_ALPHA, GL11.GL_ONE_MINUS_SRC_ALPHA, GL11.GL_ONE, GL11.GL_ZERO)
         GL11.glLineWidth(4.0f)
 
-        val x = pos.x.toDouble()
-        val y = pos.y.toDouble()
-        val z = pos.z.toDouble()
-
-        worldRenderer.begin(GL11.GL_LINES, DefaultVertexFormats.POSITION)
-        GlStateManager.color(outlineColor.red/255f, outlineColor.green/255f, outlineColor.blue/255f, outlineColor.alpha/255f)
-        worldRenderer.drawBoxOutline(Point3d(x, y, z), Point3d(x + 1, y + 1, z + 1))
-        tessellator.draw()
+        val x = pos.x.toDouble() + .333
+        val y = pos.y.toDouble() + 1
+        val z = pos.z.toDouble() + .333
 
 
         worldRenderer.begin(GL11.GL_QUADS, DefaultVertexFormats.POSITION)
         GlStateManager.color(fillColor.red/255f, fillColor.green/255f, fillColor.blue/255f, fillColor.alpha/255f)
-        worldRenderer.drawBoxFill(Point3d(x, y, z), Point3d(x + 1, y + 1, z + 1))
+
+        worldRenderer.drawBoxFill(Point3d(x, y, z), Point3d(x + .333, 256.0, z + .333))
         tessellator.draw()
 
         GlStateManager.resetColor()
         GlStateManager.enableTexture2D()
         GlStateManager.disableBlend()
         GlStateManager.enableDepth()
-
 
         GlStateManager.popMatrix()
     }
